WIP: Decouple core from DiskSpaceMonitor
Preliminary modification for JENKINS-26466. Instead of explicitly
calling the DiskSpaceMonitor, hence tightly coupling the core to it,
this commit forces a monitoring update.

As the current API only supports doing so for all the nodes, and that
could be an issue with a lot of nodes, the API is a bit enriched to be
able to selectively ask for an update for a specific computer.